+++ lib/widgets/app_bar/task_app_bar.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/tasks/linked_duration.dart';
 
@@ -38,7 +37,6 @@ class TaskAppBar extends StatelessWidget implements PreferredSizeWidget {
           return const TitleAppBar(title: 'Lotti');
         } else {
           return AppBar(
-            backgroundColor: styleConfig().cardColor,
             title: LinkedDuration(task: item),
             centerTitle: true,
             leadingWidth: 100,